+++ react-components/src/components/Header/Header.module.css
@@ -1,8 +1,9 @@
 .header {
   display: flex;
   justify-content: space-between;
+  align-items: center;
   position: sticky;
-  top: 0;
+  top: 5px;
   left: 0;
   z-index: 10;
   padding: 15px;
