+++ react-components/src/components/Header/Header.tsx
@@ -1,5 +1,6 @@
 import Pagination from '../Pagination/Pagination';
 import Search from '../Search/Search';
+import ErrorButton from '../errorBoundary/errorButton';
 import ItemsAmount from '../itemsAmount/ItemsAmount';
 import classes from './Header.module.css';
 
@@ -10,6 +11,7 @@ export default function Header() {
       <div className={classes.pageSettings}>
         <Pagination />
         <ItemsAmount />
+        <ErrorButton />
       </div>
     </header>
   );
