+++ react-components/src/components/NotFoundPage/NotFoundPage.test.tsx
@@ -1,16 +1,13 @@
 import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import { it, describe, expect } from 'vitest';
-import App from '../../App';
-import { AppProvider } from '../context/appContext';
+import { AppProvider } from '../../context/appContext';
 
 describe('NotFoundPage test', () => {
   it('Render NotFoundPage component', () => {
     render(
       <MemoryRouter initialEntries={['/non-existing-route']}>
-        <AppProvider>
-          <App />
-        </AppProvider>
+        <AppProvider>{/* <App /> */}</AppProvider>
       </MemoryRouter>
     );
     const notFoundPage = screen.getByTestId('not-found-page');
