+++ react-components/src/components/ProductDetails/ProductDetails.module.css
@@ -1,5 +1,6 @@
 .details {
   display: flex;
+  flex: 1 1 auto;
   flex-direction: column;
   width: 0;
   max-height: 800px;
