+++ react-components/src/components/RootLayout/RootLayout.module.css
@@ -1,0 +1,10 @@
+.rootLayout {
+  max-width: 1280px;
+  min-height: 100vh;
+  margin: 0 auto;
+  padding: 5px 10px;
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+  background: rgb(255, 255, 255);
+}
