+++ react-components/src/components/errorBoundary/errorButton.module.css
@@ -1,0 +1,19 @@
+.errorButton {
+  padding: 5px;
+  height: 30px;
+  border-radius: 5px;
+  border: none;
+  background: red;
+  color: white;
+  box-shadow: 2px 2px 5px 2px rgb(155, 155, 155, 0.5);
+  transition: all ease-in-out 0.2s;
+}
+
+.errorButton:hover {
+  cursor: pointer;
+  box-shadow: 2px 2px 5px 2px rgb(233, 42, 42);
+}
+
+.errorButton:active {
+  transform: scale(0.95);
+}
