+++ react-components/src/components/errorBoundary/errorButton.tsx
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import styles from './errorButton.module.css';
 
 export default function ErrorButton() {
   const [hasError, setErrorState] = useState(false);
@@ -10,7 +11,7 @@ export default function ErrorButton() {
   return (
     <button
       type="button"
-      className="error-generator"
+      className={styles.errorButton}
       onClick={() => setErrorState(true)}
     >
       Throw Error
