+++ react-components/src/pages/404.tsx
@@ -1,0 +1,3 @@
+export default function NotFoundPage() {
+  return <p data-testid="not-found-page">Such page does not exist</p>;
+}
