+++ react-components/src/state/store.ts
@@ -24,4 +24,4 @@ export type AppStore = ReturnType<typeof store>;
 export type RootState = ReturnType<AppStore['getState']>;
 export type AppDispatch = AppStore['dispatch'];
 
-export const wrapper = createWrapper<AppStore>(store, { debug: true });
+export const wrapper = createWrapper<AppStore>(store, { debug: false });
