+++ react-components/src/styles/global.css
@@ -1,0 +1,13 @@
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+.main {
+  width: 100%;
+  padding: 30px 0;
+  display: flex;
+  flex: 1 1 auto;
+  gap: 30px;
+}
